+++ lib/cext/include/ruby/intern.h
@@ -529,11 +529,13 @@ VALUE rb_hash_size(VALUE);
 void rb_hash_free(VALUE);
 /* io.c */
 #define rb_defout rb_stdout
+#ifndef TRUFFLERUBY
 RUBY_EXTERN VALUE rb_fs;
 RUBY_EXTERN VALUE rb_output_fs;
 RUBY_EXTERN VALUE rb_rs;
 RUBY_EXTERN VALUE rb_default_rs;
 RUBY_EXTERN VALUE rb_output_rs;
+#endif
 VALUE rb_io_write(VALUE, VALUE);
 VALUE rb_io_gets(VALUE);
 VALUE rb_io_getbyte(VALUE);