+++ lib/truffle/truffle/cext_constants.rb
@@ -9,24 +9,16 @@
 # From ./tool/generate-cext-constants.rb
 
 module Truffle::CExt
-  def Qundef
-    Truffle::UNDEFINED
-  end
-
-  def Qtrue
-    true
-  end
-
-  def Qfalse
-    false
-  end
-
-  def Qnil
-    nil
-  end
-
   def rb_cArray
     Array
+  end
+
+  def rb_cBasicObject
+    BasicObject
+  end
+
+  def rb_cBinding
+    Binding
   end
 
   def rb_cClass
@@ -37,16 +29,20 @@ module Truffle::CExt
     Comparable
   end
 
+  def rb_cComplex
+    Complex
+  end
+
   def rb_cData
     Data
   end
 
-  def rb_cEncoding
-    Encoding
+  def rb_cDir
+    Dir
   end
 
-  def rb_eEncodingError
-    EncodingError
+  def rb_cEncoding
+    Encoding
   end
 
   def rb_mEnumerable
@@ -65,8 +61,20 @@ module Truffle::CExt
     File
   end
 
+  def rb_mFileTest
+    FileTest
+  end
+
+  def rb_cStat
+    File::Stat
+  end
+
   def rb_cFloat
     Float
+  end
+
+  def rb_mGC
+    GC
   end
 
   def rb_cHash
@@ -89,6 +97,14 @@ module Truffle::CExt
     MatchData
   end
 
+  def rb_mMath
+    Math
+  end
+
+  def rb_cMethod
+    Method
+  end
+
   def rb_cModule
     Module
   end
@@ -105,12 +121,24 @@ module Truffle::CExt
     Object
   end
 
+  def rb_cProc
+    Proc
+  end
+
   def rb_mProcess
     Process
   end
 
+  def rb_cRandom
+    Random
+  end
+
   def rb_cRange
     Range
+  end
+
+  def rb_cRational
+    Rational
   end
 
   def rb_cRegexp
@@ -141,20 +169,16 @@ module Truffle::CExt
     TrueClass
   end
 
-  def rb_cProc
-    Proc
-  end
-
-  def rb_cMethod
-    Method
-  end
-
-  def rb_cDir
-    Dir
+  def rb_cUnboundMethod
+    UnboundMethod
   end
 
   def rb_eArgError
     ArgumentError
+  end
+
+  def rb_eEncodingError
+    EncodingError
   end
 
   def rb_eEOFError
@@ -171,6 +195,10 @@ module Truffle::CExt
 
   def rb_eFloatDomainError
     FloatDomainError
+  end
+
+  def rb_eFrozenError
+    FrozenError
   end
 
   def rb_eIndexError
@@ -249,6 +277,10 @@ module Truffle::CExt
     StandardError
   end
 
+  def rb_eStopIteration
+    StopIteration
+  end
+
   def rb_eSyntaxError
     SyntaxError
   end
@@ -287,34 +319,6 @@ module Truffle::CExt
 
   def rb_eFatal
     Truffle::CExt.rb_const_get(Object, 'fatal')
-  end
-
-  def rb_stdin
-    $stdin
-  end
-
-  def rb_stdout
-    $stdout
-  end
-
-  def rb_stderr
-    $stderr
-  end
-
-  def rb_output_fs
-    $,
-  end
-
-  def rb_rs
-    $/
-  end
-
-  def rb_output_rs
-    $\
-  end
-
-  def rb_default_rs
-    "\n"
   end
 
 end